feat : merge LocaldateUtility and fix DI issue
