+++ src/main/java/com/example/application/mfschemes/service/MFSchemeNavService.java
@@ -1,8 +1,8 @@
 package com.example.application.mfschemes.service;
 
 import com.example.application.mfschemes.NavNotFoundException;
-import com.example.application.mfschemes.util.LocalDateUtility;
 import com.example.application.mfschemes.util.SchemeConstants;
+import com.example.application.shared.LocalDateUtility;
 import com.example.application.shared.MFSchemeDTO;
 import java.time.LocalDate;
 import org.slf4j.Logger;
