+++ src/main/java/com/example/application/shared/CommonConstants.java
@@ -1,0 +1,11 @@
+package com.example.application.shared;
+
+import java.time.format.DateTimeFormatter;
+import java.util.Locale;
+
+public class CommonConstants {
+
+    private static final String DATE_PATTERN_DD_MMM_YYYY = "dd-MMM-yyyy";
+    public static final DateTimeFormatter FORMATTER_DD_MMM_YYYY =
+            DateTimeFormatter.ofPattern(DATE_PATTERN_DD_MMM_YYYY, Locale.ENGLISH);
+}
