+++ src/main/java/com/example/application/shared/MfSchemeService.java
@@ -1,5 +1,6 @@
 package com.example.application.shared;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,4 +9,8 @@ public interface MfSchemeService {
     Optional<MFSchemeProjection> findByPayOut(String isin);
 
     List<FundDetailProjection> fetchSchemes(String scheme);
+
+    Optional<MFSchemeDTO> getMfSchemeDTO(Long schemeCode, LocalDate navDate);
+
+    void fetchSchemeDetails(Long schemeCode);
 }
